% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubHits.R
\name{HubHits}
\alias{HubHits}
\alias{prop.HubHits}
\alias{print.HubHits}
\title{Quantify topic occurence in *Hub resources and combine package related
resources}
\usage{
HubHits(topics, pkg_pattern, hub = c("ExperimentHub", "AnnotationHub"))

prop.HubHits(x)

\method{print}{HubHits}(x)
}
\arguments{
\item{topics}{\code{character()} A vector of topics to query \code{ExperimentHub}
resources with}

\item{pkg_pattern}{\code{character(1)} A pattern to use for searching through
packages that expose \code{ExperimentHub} resources related to \code{topics}}

\item{hub}{\code{character(1)} Either 'ExperimentHub' or 'AnnotationHub'
indicating which hub to query resources from}

\item{x}{\code{HubHits} An instance of \code{HubHits} as returned by the eponymous
constructor function}
}
\value{
\itemize{
\item HubHits: A \code{HubHits} instance
\item prop.HubHits: The proportion of search hits to all resources in the
specified hub
}
}
\description{
Quantify topic occurence in *Hub resources and combine package related
resources
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(HubHits)}: 

}}
\section{Functions}{
\itemize{
\item \code{prop.HubHits()}: 

}}
\examples{
if (interactive()) {
et <- HubHits(c("cancer", "tumor"), "curated*", hub = "ExperimentHub")
}
if (interactive()) {
# proportion of hits in all *Hub
prop.HubHits(et)
}
if (interactive()) {
# print HubHits
et
}
}
