% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHubTopicProp.R
\name{HubTopicProp}
\alias{HubTopicProp}
\title{Quantify topic occurence in ExperimentHub resources and combine package
related resources}
\usage{
HubTopicProp(topics, pkg_pattern, hub = c("ExperimentHub", "AnnotationHub"))
}
\arguments{
\item{topics}{`character()` A vector of topics to query `ExperimentHub`
resources with}

\item{pkg_pattern}{`character(1)` A pattern to use for searching through
packages that expose `ExperimentHub` resources related to `topics`}

\item{hub}{`character(1)` Either 'ExperimentHub' or 'AnnotationHub'
indicating which hub to query resources from}
}
\value{
The proportion of all hub resources that match both the topics and
  package pattern
}
\description{
Quantify topic occurence in ExperimentHub resources and combine package
related resources
}
\examples{
et <- HubTopicProp(
    c("cancer", "tumor"), "curated*", hubtype = "AnnotationHub"
)

}
