% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubHits.R
\name{HubHits}
\alias{HubHits}
\title{Quantify topic occurence in ExperimentHub resources and combine package
related resources}
\usage{
HubHits(topics, pkg_pattern, hub = c("ExperimentHub", "AnnotationHub"))
}
\arguments{
\item{topics}{`character()` A vector of topics to query `ExperimentHub`
resources with}

\item{pkg_pattern}{`character(1)` A pattern to use for searching through
packages that expose `ExperimentHub` resources related to `topics`}

\item{hub}{`character(1)` Either 'ExperimentHub' or 'AnnotationHub'
indicating which hub to query resources from}
}
\value{
The proportion of all hub resources that match both the topics and
  package pattern
}
\description{
Quantify topic occurence in ExperimentHub resources and combine package
related resources
}
\examples{
et <- HubHits(
    c("cancer", "tumor"), "curated*", hub = "ExperimentHub"
)
# proportion
et[1] / et[2]

}
